% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignExtraCluster.R
\name{assignExtraCluster}
\alias{assignExtraCluster}
\title{Assign cells in the _Extra cluster based on their UMAP coordinates}
\usage{
assignExtraCluster(
  SeuData,
  ClusterName,
  min.x = -100,
  max.x = 100,
  min.y = -100,
  max.y = 100
)
}
\arguments{
\item{SeuData}{Seurat object with UMAP coordinates}

\item{ClusterName}{Name of the cluster to assign the _Extra cells to}

\item{min.x}{Minimum x value for the _Extra cluster boundary}

\item{max.x}{Maximum x value for the _Extra cluster boundary}

\item{min.y}{Minimum y value for the _Extra cluster boundary}

\item{max.y}{Maximum y value for the _Extra cluster boundary}
}
\value{
A Seurat object with the _Extra cluster assigned
}
\description{
This function assigns cells in the _Extra cluster based on their UMAP coordinates.
}
\examples{
assignExtraCluster(SeuratObject, ClusterName = "Cluster1", -10, 10, -10, 10)
}
