% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignNewCluster.R
\name{assignNewCluster}
\alias{assignNewCluster}
\title{assignNewCluster function}
\usage{
assignNewCluster(
  SeuData,
  ClusterName,
  min.x = -100,
  max.x = 100,
  min.y = -100,
  max.y = 100,
  validate = FALSE
)
}
\arguments{
\item{SeuData}{A Seurat object that contains the data of the clusters.}

\item{ClusterName}{Name of the new cluster to be generated.}

\item{min.x}{Minimum x coordinate of the new cluster boundaries. The default is `-100`.}

\item{max.x}{Maximum x coordinate of the new cluster boundaries. The default is `100`.}

\item{min.y}{Minimum y coordinate of the new cluster boundaries. The default is `-100`.}

\item{max.y}{Maximum y coordinate of the new cluster boundaries. The default is `100`.}

\item{validate}{Boolean value indicating whether to validate the assignment. This must be set to `TRUE`
for the new cells to be successfully assigned and subtracted from the unassigned cells.
The default is `FALSE`.}
}
\value{
Invisibly returns a Seurat UMAP plot of all clusters in the data. This function also updates the
        global list "datacleaning_list" by adding a new cluster and updating the leftovers.
}
\description{
This function selects cells from the leftovers (cells not assigned to any cluster yet) based on
provided x and y coordinates, assigns them to a new cluster, and updates the leftovers.
In order for the new cells to be successfully assigned and subtracted from the unassigned cells,
the validate parameter must be set to TRUE.
}
