% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckUMAP1.R
\name{CheckUMAP1}
\alias{CheckUMAP1}
\title{Check UMAP plot for a specific subset of cells}
\usage{
CheckUMAP1(CheckInput, sample = SeuData)
}
\arguments{
\item{CheckInput}{A data frame or matrix with UMAP coordinates to check}

\item{sample}{A Seurat object with UMAP coordinates}
}
\value{
A UMAP plot for the specified subset of cells
}
\description{
This function checks the UMAP plot for a specific subset of cells.
}
