% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citibi_raw.R
\docType{data}
\name{citbi_raw}
\alias{citbi_raw}
\title{A tibble of a subset of the Pediatric Clinically
Important Traumatic Brain Injury (ciTBI) dataset}
\format{
A tibble with 30,379 rows and 26 columns. 
Each row represents a different child while each column
is a different measure after examination of the child.
}
\source{
Stat 133
}
\usage{
citbi_raw
}
\description{
This data set contains information on children who
visited the emergency department with head trauma.
}
\keyword{datasets}
