% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropCluster.R
\name{cropCluster}
\alias{cropCluster}
\title{cropCluster function}
\usage{
cropCluster(
  SeuData,
  ClusterName,
  ClusterNo = 0,
  min.x = -100,
  max.x = 100,
  min.y = -100,
  max.y = 100
)
}
\arguments{
\item{SeuData}{A Seurat object that contains the data of the clusters.}

\item{ClusterName}{Name of the cluster to be selected.}

\item{ClusterNo}{Number of the cluster to be selected. The default is `0`.}

\item{min.x}{Minimum x coordinate of the new cluster boundaries. The default is `-100`.}

\item{max.x}{Maximum x coordinate of the new cluster boundaries. The default is `100`.}

\item{min.y}{Minimum y coordinate of the new cluster boundaries. The default is `-100`.}

\item{max.y}{Maximum y coordinate of the new cluster boundaries. The default is `100`.}
}
\value{
Invisibly returns a Seurat UMAP plot of all clusters in the data.
       This function also generates two global variables, one for the cleaned cluster
       and one for the leftover cells, and adds them to the global list "datacleaning_list".
}
\description{
This function selects a cluster from the Seurat object based on the provided cluster name or number,
adjusts its boundaries based on provided min and max coordinates in both axes, and then stores
the cleaned cluster and leftovers in datacleaning_list in the global environment.
}
