\name{datacleaning1-package}
\alias{datacleaning1-package}
\alias{datacleaning1}
\docType{package}
\title{
\packageTitle{datacleaning1}
}
\description{
\packageDescription{datacleaning1}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{datacleaning1}

Maintainer: \packageMaintainer{datacleaning1}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = datacleaning1)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{datacleaning1}}
%\section{Documentation Index}{\packageIndices{datacleaning1}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
