% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{forest_logic}
\alias{forest_logic}
\title{Analyse input and create model}
\usage{
forest_logic(input)
}
\arguments{
\item{input}{the user input}
}
\value{
predict_price
the predicted sale price of the input
}
\description{
The function will read the input that was taken in from the user.
It will convert factored/ordered inputs into numerics to put into the prediction model
It then puts these inputs into a dataframe (with a single row).
This single row is then plotted against the random forest model in order to predict its sale price
}
