% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFinalClusters.R
\name{getFinalClusters}
\alias{getFinalClusters}
\title{getFinalClusters function}
\usage{
getFinalClusters(SeuData)
}
\arguments{
\item{SeuData}{A Seurat object that needs to have its cell identities updated.}
}
\value{
SeuData The updated Seurat object, with the new cell identities reflected in the metadata.
It also prints a UMAP plot of the updated Seurat object.
}
\description{
This function updates the identities of the cells in the input Seurat object
according to the "_Clean" and "_Extra" clusters defined in the global "datacleaning_list",
and plots the final UMAP visualization. It also updates the "celltype" column in
the Seurat object's metadata and removes the "datacleaning_list" from the global environment.
}
\examples{
# Assuming 'SeuratObject' is your Seurat object:
SeuratObject <- getFinalClusters(SeuratObject)
}
