% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLeftovers.R
\name{getLeftovers}
\alias{getLeftovers}
\title{Get Leftover cells not assigned to any clusters}
\usage{
getLeftovers(SeuData)
}
\arguments{
\item{SeuData}{A Seurat object containing the UMAP coordinates of each cell.}
}
\value{
A data frame containing the UMAP coordinates of each leftover cell.
}
\description{
This function gets the UMAP coordinates of each cell and retrieves the cell barcodes in the Seurat object that are not assigned to any clusters. It then saves the Leftovers object in the global environment and returns the result.
}
\examples{
getLeftovers(SeuratObject)
}
