% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_query.R
\name{queryLoad}
\alias{queryLoad}
\alias{querySave}
\title{Load and Save the queries from an excel workbook.}
\usage{
queryLoad(file, queryRun = 1)

querySave(wb, q, file)
}
\arguments{
\item{file}{fileName to load from or save to}

\item{queryRun}{The number of the current}

\item{wb}{The workbook object created by queryLoad}

\item{q}{The query object with all queries checked and updated}
}
\description{
Load and Save the queries from an excel workbook.
}
\details{
As a minimum the file to load should contain an "Instructions" and an "Initials" worksheet.
The database has 12 columns: "id","identifier","queryRun","patid","CRF","repeatLine1","repeatLine2","firstQuery","STATScomments","comments","resolved","STATSresolved" these are checked by queryLoad to ensure there will be no errors later.
}
\section{Functions}{
\itemize{
\item \code{querySave}: Save the query object to excel
}}

