% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateExtraCluster.R
\name{validateExtraCluster}
\alias{validateExtraCluster}
\title{Validate and visualize the updated Leftovers after assigning Extra cluster}
\usage{
validateExtraCluster(SeuData)
}
\arguments{
\item{SeuData}{Seurat object containing the data to be analyzed}
}
\value{
A plot of the UMAP coordinates of the updated Leftovers
}
\description{
Validate and visualize the updated Leftovers after assigning Extra cluster
}
\examples{
validateExtraCluster(SeuratObject)
}
