% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUMAP.R
\name{CheckUMAP}
\alias{CheckUMAP}
\title{CheckUMAP function}
\usage{
CheckUMAP(CheckInput, sample, color = "#cb181d", label.text = "RoI")
}
\arguments{
\item{CheckInput}{A data.frame or object to be checked. If it is a data.frame,
the function generates metadata based on the row names.
If it is not a data.frame, the function generates metadata
based on the column names.}

\item{sample}{A Seurat object with UMAP coordinates. The default is `SeuData`.}

\item{color}{The color for the UMAP plot. The default is `#cb181d`.}

\item{label.text}{The text to use in generating the metadata. The default is `"RoI"`.}
}
\value{
A UMAP plot for the specified subset of cells
}
\description{
This function adds metadata to the input Seurat object, assigns the metadata to sample's Idents,
and then generates a UMAP plot of the Seurat object with highlighted cells of selected cluster.
}
