% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFinalClusters.R
\name{getFinalClusters}
\alias{getFinalClusters}
\title{Get Final Clusters}
\usage{
getFinalClusters(SeuData)
}
\arguments{
\item{SeuData}{Seurat object containing the data}
}
\value{
Updated Seurat object with the Idents and celltype columns updated
}
\description{
This function updates the Idents of the clusters in the Seurat object based on the object names ending with _Clean or _Extra.
}
\examples{
getFinalClusters(SeuData)
}
