% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_query.R
\name{queryLoad}
\alias{queryLoad}
\alias{querySave}
\title{Load and Save the queries from an excel workbook.}
\usage{
queryLoad(file, queryRun = 1)

querySave(wb, q, file)
}
\arguments{
\item{file}{fileName to load from or save to}

\item{queryRun}{The number of the current}

\item{wb}{The workbook object created by queryLoad}

\item{q}{The query object with all queries checked and updated}
}
\description{
Load and Save the queries from an excel workbook.
}
\section{Functions}{
\itemize{
\item \code{querySave}: Save the query object to excel
}}

